chore: reparent aibridge code under enterprise
