+++ enterprise/cli/server.go
@@ -7,6 +7,7 @@ import (
 	"database/sql"
 	"encoding/base64"
 	"errors"
+	"fmt"
 	"io"
 	"net/url"
 
@@ -15,6 +16,7 @@ import (
 	"tailscale.com/types/key"
 
 	"github.com/coder/coder/v2/coderd/database"
+	"github.com/coder/coder/v2/codersdk"
 	"github.com/coder/coder/v2/cryptorand"
 	"github.com/coder/coder/v2/enterprise/audit"
 	"github.com/coder/coder/v2/enterprise/audit/backends"
@@ -23,6 +25,7 @@ import (
 	"github.com/coder/coder/v2/enterprise/coderd/usage"
 	"github.com/coder/coder/v2/enterprise/dbcrypt"
 	"github.com/coder/coder/v2/enterprise/trialer"
+	"github.com/coder/coder/v2/enterprise/x/aibridged"
 	"github.com/coder/coder/v2/tailnet"
 	"github.com/coder/quartz"
 	"github.com/coder/serpent"
@@ -142,6 +145,33 @@ func (r *RootCmd) Server(_ func()) *serpent.Command {
 			return nil, nil, xerrors.Errorf("start usage publisher: %w", err)
 		}
 		closers.Add(publisher)
+
+		experiments := agplcoderd.ReadExperiments(options.Logger, options.DeploymentValues.Experiments.Value())
+
+		var aibridgeDaemon *aibridged.Server
+		// In-memory aibridge daemon.
+		if options.DeploymentValues.AI.BridgeConfig.Enabled {
+			if experiments.Enabled(codersdk.ExperimentAIBridge) {
+				aibridgeDaemon, err = newAIBridgeDaemon(api)
+				if err != nil {
+					return nil, nil, xerrors.Errorf("create aibridged: %w", err)
+				}
+
+				api.RegisterInMemoryAIBridgedHTTPHandler(aibridgeDaemon)
+
+				// When running as an in-memory daemon, the HTTP handler is wired into the
+				// coderd API and therefore is subject to its context. Calling Close() on
+				// aibridged will NOT affect in-flight requests but those will be closed once
+				// the API server is itself shutdown.
+				closers.Add(aibridgeDaemon)
+			} else {
+				api.Logger.Warn(ctx, fmt.Sprintf("CODER_AIBRIDGE_ENABLED=true but experiment %q not enabled", codersdk.ExperimentAIBridge))
+			}
+		} else {
+			if experiments.Enabled(codersdk.ExperimentAIBridge) {
+				api.Logger.Warn(ctx, "aibridge experiment enabled but CODER_AIBRIDGE_ENABLED=false")
+			}
+		}
 
 		return api.AGPL, closers, nil
 	})
